# src/backend/api/docs/swagger.yaml
basePath: /api
definitions:
...
  model.ChatMessage:
    properties:
      chat_session_id:
        type: number
      citations:
        items:
          $ref: '#/definitions/model.DocumentResponse'
        type: array
      error:
        type: string
      feedback:
        $ref: '#/definitions/model.ChatMessageFeedback'
      id:
        type: number
      latest_child_message:
        type: integer
      message:
        type: string
      message_type:
        type: string
      parent_message:
        type: number
      rephrased_query:
        type: string
      time_sent:
        type: string
      token_count:
        type: integer
    type: object
  model.ChatMessageFeedback:
    properties:
      chat_message_id:
        type: number
      feedback:
        type: string
      id:
        type: number
      up_votes:
        type: boolean
      user_id:
        type: string
    type: object
  model.ChatSession:
    properties:
      creation_date:
        type: string
      deleted_date:
        $ref: '#/definitions/pg.NullTime'
      description:
        type: string
      id:
        type: number
      messages:
        items:
          $ref: '#/definitions/model.ChatMessage'
        type: array
      one_shot:
        type: boolean
      persona:
        $ref: '#/definitions/model.Persona'
      persona_id:
        type: number
      user_id:
        type: string
    type: object
...
  model.DocumentResponse:
    properties:
      content:
        type: string
      document_id:
        type: string
      id:
        type: number
      link:
        type: string
      message_id:
        type: number
      updated_date:
        type: string
    type: object
...
  model.JSONMap:
    additionalProperties: true
    type: object
  model.LLM:
    properties:
      api_key:
        type: string
      creation_date:
        type: string
      deleted_date:
        $ref: '#/definitions/pg.NullTime'
      endpoint:
        type: string
      id:
        type: number
      last_update:
        $ref: '#/definitions/pg.NullTime'
      model_id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
      url:
        type: string
    type: object
  model.Persona:
    properties:
      chat_sessions:
        items:
          $ref: '#/definitions/model.ChatSession'
        type: array
      creation_date:
        type: string
      default_persona:
        type: boolean
      deleted_date:
        $ref: '#/definitions/pg.NullTime'
      description:
        type: string
      display_priority:
        type: integer
      id:
        type: number
      is_visible:
        type: boolean
      last_update:
        $ref: '#/definitions/pg.NullTime'
      llm:
        $ref: '#/definitions/model.LLM'
      llm_id:
        type: number
      name:
        type: string
      prompt:
        $ref: '#/definitions/model.Prompt'
      starter_messages:
        items:
          type: integer
        type: array
      tenant_id:
        type: string
    type: object
  model.Prompt:
    properties:
      creation_date:
        type: string
      deleted_date:
        $ref: '#/definitions/pg.NullTime'
      description:
        type: string
      id:
        type: number
      last_update:
        $ref: '#/definitions/pg.NullTime'
      name:
        type: string
      persona_id:
        type: number
      system_prompt:
        type: string
      task_prompt:
        type: string
      user_id:
        type: string
    type: object
...
  model.Tenant:
    properties:
      configuration:
        $ref: '#/definitions/model.JSONMap'
      id:
        type: string
      name:
        type: string
    type: object
  model.User:
    properties:
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      roles:
        items:
          type: string
        type: array
      tenant:
        $ref: '#/definitions/model.Tenant'
      tenant_id:
        type: string
      user_name:
        type: string
    type: object
  parameters.AddUserParam:
    properties:
      email:
        type: string
      role:
        type: string
    type: object
...
  parameters.EditUserParam:
    properties:
      role:
        type: string
    type: object
...
  parameters.PersonaParam:
    properties:
      api_key:
        type: string
      description:
        type: string
      endpoint:
        type: string
      model_id:
        type: string
      name:
        type: string
      starter_messages:
        items:
          $ref: '#/definitions/parameters.StarterMessage'
        type: array
      system_prompt:
        type: string
      task_prompt:
        type: string
      url:
        type: string
    type: object
  parameters.StarterMessage:
    properties:
      description:
        type: string
      message:
        type: string
      name:
        type: string
    type: object
...
  pg.NullTime:
    properties:
      time.Time:
        type: string
    type: object
...
info:
  contact:
    name: API Support
  description: This is Cognix Golang API Documentation
  title: Cognix API
  version: "1.0"
paths:
...
  /manage/personas:
    get:
      description: return list of allowed personas
      operationId: personas_get_all
      parameters:
      - description: true for include deleted personas
        in: query
        name: archived
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Persona'
            type: array
      security:
      - ApiKeyAuth: []
      summary: return list of allowed personas
      tags:
      - Persona
    post:
      description: create persona
      operationId: persona_create
      parameters:
      - description: persona payload
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/parameters.PersonaParam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Persona'
      security:
      - ApiKeyAuth: []
      summary: create persona
      tags:
      - Persona
  /manage/personas/{id}:
    get:
      description: return persona by id
      operationId: persona_get_by_id
      parameters:
      - description: persona id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Persona'
      security:
      - ApiKeyAuth: []
      summary: return persona by id
      tags:
      - Persona
    put:
      description: update persona
      operationId: persona_update
      parameters:
      - description: persona id
        in: path
        name: id
        required: true
        type: integer
      - description: persona payload
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/parameters.PersonaParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Persona'
      security:
      - ApiKeyAuth: []
      summary: update persona
      tags:
      - Persona
  /manage/personas/{id}/{action}:
    post:
      description: delete or restore persona
      operationId: persona_delete_restore
      parameters:
      - description: persona id
        in: path
        name: id
        required: true
        type: integer
      - description: 'action : delete | restore '
        in: path
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Persona'
      security:
      - ApiKeyAuth: []
      summary: delete or restore persona
      tags:
      - Persona
  /tenant/user_info:
    get:
      description: get user info
      operationId: tenant_get_user_info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      security:
      - ApiKeyAuth: []
      summary: get user info
      tags:
      - Tenant
  /tenant/users:
    get:
      description: return list of users
      operationId: tenant_get_users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.User'
            type: array
      security:
      - ApiKeyAuth: []
      summary: return list of users
      tags:
      - Tenant
    post:
      description: add new user
      operationId: tenant_add_user
      parameters:
      - description: create user parameter
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/parameters.AddUserParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      security:
      - ApiKeyAuth: []
      summary: add new user
      tags:
      - Tenant
  /tenant/users/{id}:
    put:
      description: edit  user
      operationId: tenant_edit_user
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: edit user parameter
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/parameters.EditUserParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      security:
      - ApiKeyAuth: []
      summary: edit user
      tags:
      - Tenant
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"